[CHORE] #90 Fix data struct file
